Fix 32 bit windows build

+++ custom/4coder_file.cpp
@@ -19,7 +19,7 @@ data_from_file(Arena *arena, FILE *file){
         result.size = ftell(file);
         fseek(file, 0, SEEK_SET);
         result.str = push_array(arena, u8, result.size + 1);
-        fread(result.str, 1, result.size, file);
+        fread(result.str, 1, (size_t)result.size, file);
         result.str[result.size] = 0;
     }
     return(result);
